Consolidate envelope UGens
